requirements.txt
python-dotenv==0.20.0
requests==2.28.1
dependency-injector=4.40.0
